--- readme.md
@@ -1,1 +1,0 @@
-## Scany
